move the score updating logic into a new class and we now use a separated thread
