+++ app/src/main/java/ch/epfl/sdp/game/GameThread.java
@@ -95,12 +95,4 @@ public class GameThread extends Thread{
             System.out.println("Average FPS: " + avgFPS);
         }
     }
-
-    public long getTotalTime() {
-        return totalTime;
-    }
-
-    public int getFPS() {
-        return FPS;
-    }
 }
