+++ app/src/test/java/ch/epfl/sdp/entity/PlayerTest.java
@@ -5,6 +5,7 @@ import org.junit.Test;
 
 import ch.epfl.sdp.game.Game;
 import ch.epfl.sdp.item.Healthpack;
+import ch.epfl.sdp.map.MockMapApi;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -18,7 +19,7 @@ public class PlayerTest {
     public void setup(){
         PlayerManager playerManager = new PlayerManager();
         player1 = new Player(6.149290, 46.212470, 50, "Skyris", "test@email.com");
-
+        game.getInstance().setMapApi(new MockMapApi());
         PlayerManager.setCurrentUser(player1);
     }
